LPE

Description

Documentation

type ProcDefs = Map ProcId ProcDef

preGNF :: EnvB envb => ProcId -> TranslatedProcDefs -> ProcDefs -> envb ProcDefs

gnf :: EnvB envb => ProcId -> TranslatedProcDefs -> ProcDefs -> envb ProcDefs

lpeTransform

Arguments

:: EnvB envb

=> BExpr
behaviour expression to be transformed, assumed to be a process instantiation

-> ProcDefs
context of process definitions in which process instantiation is assumed to be defined

-> envb (Maybe (BExpr, ProcDef))
transformed process instantiation with LPE definition template function for lpe

wrapper around lpe function, returning only the relevant ProcDef instead of all ProcDefs

lpe :: EnvB envb => BExpr -> TranslatedProcDefs -> ProcDefs -> envb (BExpr, ProcDefs)

lpePar :: EnvB envb => BExpr -> TranslatedProcDefs -> ProcDefs -> envb (BExpr, ProcDefs)

lpeHide :: EnvB envb => BExpr -> TranslatedProcDefs -> ProcDefs -> envb (BExpr, ProcDefs)

preGNFEnable :: EnvB envb => BExpr -> TranslatedProcDefs -> ProcDefs -> envb (BExpr, ProcDefs)